Move some shared code into an internal package
