+++ kubectl-fdb/cmd/deprecation_test.go
@@ -24,7 +24,8 @@ import (
 	"bytes"
 	"strings"
 
-	"github.com/FoundationDB/fdb-kubernetes-operator/controllers"
+	"github.com/FoundationDB/fdb-kubernetes-operator/internal"
+
 	"github.com/spf13/cobra"
 	corev1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/resource"
@@ -49,7 +50,7 @@ var _ = Describe("[plugin] deprecation command", func() {
 		type testCase struct {
 			cluster            fdbtypes.FoundationDBCluster
 			inputClusters      []string
-			deprecationOptions controllers.DeprecationOptions
+			deprecationOptions internal.DeprecationOptions
 			showClusterSpec    bool
 			checkOutput        bool
 			expectedOutput     string
@@ -94,7 +95,7 @@ var _ = Describe("[plugin] deprecation command", func() {
 						Spec: fdbtypes.FoundationDBClusterSpec{},
 					},
 					inputClusters: []string{"no-match"},
-					deprecationOptions: controllers.DeprecationOptions{
+					deprecationOptions: internal.DeprecationOptions{
 						UseFutureDefaults: false,
 						OnlyShowChanges:   false,
 					},
@@ -148,7 +149,7 @@ var _ = Describe("[plugin] deprecation command", func() {
 						},
 					},
 					inputClusters: []string{},
-					deprecationOptions: controllers.DeprecationOptions{
+					deprecationOptions: internal.DeprecationOptions{
 						UseFutureDefaults: false,
 						OnlyShowChanges:   true,
 					},
@@ -203,7 +204,7 @@ var _ = Describe("[plugin] deprecation command", func() {
 						},
 					},
 					inputClusters: []string{},
-					deprecationOptions: controllers.DeprecationOptions{
+					deprecationOptions: internal.DeprecationOptions{
 						UseFutureDefaults: false,
 						OnlyShowChanges:   false,
 					},
@@ -258,7 +259,7 @@ var _ = Describe("[plugin] deprecation command", func() {
 						},
 					},
 					inputClusters: []string{},
-					deprecationOptions: controllers.DeprecationOptions{
+					deprecationOptions: internal.DeprecationOptions{
 						UseFutureDefaults: false,
 						OnlyShowChanges:   true,
 					},
@@ -312,7 +313,7 @@ var _ = Describe("[plugin] deprecation command", func() {
 						},
 					},
 					inputClusters: []string{},
-					deprecationOptions: controllers.DeprecationOptions{
+					deprecationOptions: internal.DeprecationOptions{
 						UseFutureDefaults: false,
 						OnlyShowChanges:   true,
 					},
@@ -366,7 +367,7 @@ var _ = Describe("[plugin] deprecation command", func() {
 						},
 					},
 					inputClusters: []string{},
-					deprecationOptions: controllers.DeprecationOptions{
+					deprecationOptions: internal.DeprecationOptions{
 						UseFutureDefaults: false,
 						OnlyShowChanges:   false,
 					},
@@ -465,7 +466,7 @@ version: ""`,
 						},
 					},
 					inputClusters: []string{},
-					deprecationOptions: controllers.DeprecationOptions{
+					deprecationOptions: internal.DeprecationOptions{
 						UseFutureDefaults: false,
 						OnlyShowChanges:   true,
 					},
